+++ src/sdam/topology.ts
@@ -50,7 +50,7 @@ import type { MongoCredentials } from '../cmap/auth/mongo_credentials';
 import type { Transaction } from '../transactions';
 import type { CloseOptions } from '../cmap/connection_pool';
 import { DestroyOptions, Connection } from '../cmap/connection';
-import type { MongoClientOptions } from '../mongo_client';
+import type { MongoClientOptions, ServerApi } from '../mongo_client';
 import { DEFAULT_OPTIONS } from '../connection_string';
 import { serialize, deserialize } from '../bson';
 
@@ -170,6 +170,8 @@ export class Topology extends EventEmitter {
   ismaster?: Document;
   /** @internal */
   _type?: string;
+  /** @internal */
+  serverApi?: ServerApi;
 
   /** @event */
   static readonly SERVER_OPENING = 'serverOpening' as const;
@@ -202,8 +204,14 @@ export class Topology extends EventEmitter {
   /**
    * @param seedlist - a list of HostAddress instances to connect to
    */
-  constructor(seeds: string | string[] | HostAddress | HostAddress[], options: TopologyOptions) {
+  constructor(
+    seeds: string | string[] | HostAddress | HostAddress[],
+    options: TopologyOptions,
+    serverApi?: ServerApi
+  ) {
     super();
+
+    this.serverApi = serverApi;
 
     // Legacy CSFLE support
     this.bson = Object.create(null);
