+++ test/functional/unified-spec-runner/runner.ts
@@ -1,0 +1,142 @@
+import { expect } from 'chai';
+import { satisfies as semverSatisfies } from 'semver';
+import * as uni from './schema';
+import { ReadPreference } from '../../../src/read_preference';
+import { zip, topologySatisfies, patchVersion } from './unified-utils';
+import { CommandEvent, EntitiesMap } from './entities';
+import { ns } from '../../../src/utils';
+import { executeOperationAndCheck } from './operations';
+import { matchesEvents } from './match';
+
+export type TestConfiguration = InstanceType<
+  typeof import('../../tools/runner/config')['TestConfiguration']
+>;
+interface MongoDBMochaTestContext extends Mocha.Context {
+  configuration: TestConfiguration;
+}
+
+export async function runUnifiedTest(
+  ctx: MongoDBMochaTestContext,
+  unifiedSuite: uni.UnifiedSuite,
+  test: uni.Test
+): Promise<void> {
+  // Some basic expectations we can catch early
+  expect(test).to.exist;
+  expect(unifiedSuite).to.exist;
+  expect(ctx).to.exist;
+  expect(ctx.configuration).to.exist;
+
+  const schemaVersion = patchVersion(unifiedSuite.schemaVersion);
+  expect(semverSatisfies(schemaVersion, uni.SupportedVersion)).to.be.true;
+
+  // If test.skipReason is specified, the test runner MUST skip this
+  // test and MAY use the string value to log a message.
+  if (test.skipReason) {
+    console.warn(`Skipping test ${test.description}: ${test.skipReason}.`);
+    ctx.skip();
+  }
+
+  const UTIL_CLIENT = ctx.configuration.newClient();
+  await UTIL_CLIENT.connect();
+  ctx.defer(async () => await UTIL_CLIENT.close());
+
+  // Must fetch parameters before checking runOnRequirements
+  ctx.configuration.parameters = await UTIL_CLIENT.db().admin().command({ getParameter: '*' });
+
+  // If test.runOnRequirements is specified, the test runner MUST skip the test unless one or more
+  // runOnRequirement objects are satisfied.
+  const allRequirements = [
+    ...(unifiedSuite.runOnRequirements ?? []),
+    ...(test.runOnRequirements ?? [])
+  ];
+  for (const requirement of allRequirements) {
+    if (!topologySatisfies(ctx.configuration, requirement)) {
+      ctx.skip();
+    }
+  }
+
+  // If initialData is specified, for each collectionData therein the test runner MUST drop the
+  // collection and insert the specified documents (if any) using a "majority" write concern. If no
+  // documents are specified, the test runner MUST create the collection with a "majority" write concern.
+  // The test runner MUST use the internal MongoClient for these operations.
+  if (unifiedSuite.initialData) {
+    for (const collData of unifiedSuite.initialData) {
+      const db = UTIL_CLIENT.db(collData.databaseName);
+      const collection = db.collection(collData.collectionName, {
+        writeConcern: { w: 'majority' }
+      });
+      const collectionList = await db.listCollections({ name: collData.collectionName }).toArray();
+      if (collectionList.length !== 0) {
+        expect(await collection.drop()).to.be.true;
+      }
+
+      if (collData.documents.length === 0) {
+        await db.createCollection(collData.collectionName, {
+          writeConcern: { w: 'majority' }
+        });
+        continue;
+      }
+
+      await collection.insertMany(collData.documents);
+    }
+  }
+
+  const entities = await EntitiesMap.createEntities(ctx.configuration, unifiedSuite.createEntities);
+  ctx.defer(async () => await entities.cleanup());
+
+  // Workaround for SERVER-39704:
+  // test runners MUST execute a non-transactional distinct command on
+  // each mongos server before running any test that might execute distinct within a transaction.
+  // To ease the implementation, test runners MAY execute distinct before every test.
+  if (
+    ctx.topologyType === uni.TopologyType.sharded ||
+    ctx.topologyType === uni.TopologyType.shardedReplicaset
+  ) {
+    for (const [, collection] of entities.mapOf('collection')) {
+      await UTIL_CLIENT.db(ns(collection.namespace).db).command({
+        distinct: collection.collectionName,
+        key: '_id'
+      });
+    }
+  }
+
+  for (const operation of test.operations) {
+    await executeOperationAndCheck(operation, entities);
+  }
+
+  const clientEvents = new Map<string, CommandEvent[]>();
+  // If any event listeners were enabled on any client entities,
+  // the test runner MUST now disable those event listeners.
+  for (const [id, client] of entities.mapOf('client')) {
+    clientEvents.set(id, client.stopCapturingEvents());
+  }
+
+  if (test.expectEvents) {
+    for (const expectedEventList of test.expectEvents) {
+      const clientId = expectedEventList.client;
+      const actualEvents = clientEvents.get(clientId);
+
+      expect(actualEvents, `No client entity found with id ${clientId}`).to.exist;
+      matchesEvents(expectedEventList.events, actualEvents, entities);
+    }
+  }
+
+  if (test.outcome) {
+    for (const collectionData of test.outcome) {
+      const collection = UTIL_CLIENT.db(collectionData.databaseName).collection(
+        collectionData.collectionName
+      );
+      const findOpts = {
+        readConcern: 'local' as const,
+        readPreference: ReadPreference.primary,
+        sort: { _id: 'asc' as const }
+      };
+      const documents = await collection.find({}, findOpts).toArray();
+
+      expect(documents).to.have.lengthOf(collectionData.documents.length);
+      for (const [expected, actual] of zip(collectionData.documents, documents)) {
+        expect(actual).to.include(expected, 'Test outcome did not match expected');
+      }
+    }
+  }
+}
