+++ test/spec/versioned-api/runcommand-helper-no-api-version-declared.json
@@ -1,0 +1,117 @@
+{
+  "description": "RunCommand helper: No API version declared",
+  "schemaVersion": "1.1",
+  "runOnRequirements": [
+    {
+      "minServerVersion": "4.7",
+      "serverParameters": {
+        "requireApiVersion": false
+      }
+    }
+  ],
+  "createEntities": [
+    {
+      "client": {
+        "id": "client",
+        "observeEvents": [
+          "commandStartedEvent"
+        ]
+      }
+    },
+    {
+      "database": {
+        "id": "database",
+        "client": "client",
+        "databaseName": "versioned-api-tests"
+      }
+    }
+  ],
+  "tests": [
+    {
+      "description": "runCommand does not inspect or change the command document",
+      "operations": [
+        {
+          "name": "runCommand",
+          "object": "database",
+          "arguments": {
+            "commandName": "ping",
+            "command": {
+              "ping": 1,
+              "apiVersion": "server_will_never_support_this_api_version"
+            }
+          },
+          "expectError": {
+            "isError": true,
+            "isClientError": false
+          }
+        }
+      ],
+      "expectEvents": [
+        {
+          "client": "client",
+          "events": [
+            {
+              "commandStartedEvent": {
+                "command": {
+                  "ping": 1,
+                  "apiVersion": "server_will_never_support_this_api_version",
+                  "apiStrict": {
+                    "$$exists": false
+                  },
+                  "apiDeprecationErrors": {
+                    "$$exists": false
+                  }
+                },
+                "commandName": "ping",
+                "databaseName": "versioned-api-tests"
+              }
+            }
+          ]
+        }
+      ]
+    },
+    {
+      "description": "runCommand does not prevent sending invalid API version declarations",
+      "operations": [
+        {
+          "name": "runCommand",
+          "object": "database",
+          "arguments": {
+            "commandName": "ping",
+            "command": {
+              "ping": 1,
+              "apiStrict": true
+            }
+          },
+          "expectError": {
+            "isError": true,
+            "isClientError": false
+          }
+        }
+      ],
+      "expectEvents": [
+        {
+          "client": "client",
+          "events": [
+            {
+              "commandStartedEvent": {
+                "command": {
+                  "ping": 1,
+                  "apiVersion": {
+                    "$$exists": false
+                  },
+                  "apiStrict": true,
+                  "apiDeprecationErrors": {
+                    "$$exists": false
+                  }
+                },
+                "commandName": "ping",
+                "databaseName": "versioned-api-tests"
+              }
+            }
+          ]
+        }
+      ]
+    }
+  ]
+}
